+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - The `map` mutation.
 - The `makeCallbackMutation` mutation builder.
 - The `cycle` mutation.
+- The `direct` mutation.
 
 ## [0.1.1] - 2018-09-05
 ### Changed